Updated mixmod code to use floats and ints

+++ similarity/clustering/MixtureModelPWClusters.cpp
@@ -27,9 +27,9 @@ MixtureModelPWClusters::MixtureModelPWClusters(PairWiseSet *pwset, int min_obs,
   }
 
   // Create the Gaussian Data object and set the dataDescription object.
-  this->data = (double **) malloc(sizeof(double *) * pwset->n_clean);
+  this->data = (float **) malloc(sizeof(float *) * pwset->n_clean);
   for (int i = 0; i < pwset->n_clean ; i++) {
-    this->data[i] = (double *) malloc(sizeof(double) * 2);
+    this->data[i] = (float *) malloc(sizeof(float) * 2);
     this->data[i][0] = pwset->x_clean[i];
     this->data[i][1] = pwset->y_clean[i];
   }
@@ -57,7 +57,7 @@ MixtureModelPWClusters::~MixtureModelPWClusters() {
  * Executes mixture model clustering.
  */
 void MixtureModelPWClusters::run(char * criterion, int max_clusters) {
-  int64_t lowest_cluster_num = 9;
+  int lowest_cluster_num = 9;
   int found = 0;
 
   // Make sure we have the correct number of observations before performing
@@ -71,7 +71,7 @@ void MixtureModelPWClusters::run(char * criterion, int max_clusters) {
   XEM::DataDescription dataDescription(gdata);
 
   // Set the number of clusters to be tested to 9.
-  vector<int64_t> nbCluster;
+  vector<int> nbCluster;
   for (int i = 1; i <= max_clusters; i++) {
     nbCluster.push_back(i);
   }
@@ -141,10 +141,10 @@ void MixtureModelPWClusters::run(char * criterion, int max_clusters) {
       XEM::ClusteringModelOutput* cMOutput = cOutput->getClusteringModelOutput().front();
       XEM::LabelDescription * ldescription = cMOutput->getLabelDescription();
       XEM::Label * label = ldescription->getLabel();
-      int64_t * cLabels = label->getTabLabel();
+      int * cLabels = label->getTabLabel();
 
       // Find the number of clusters
-      int64_t num_clusters = 0;
+      int num_clusters = 0;
       for (int i = 0; i < this->pwset->n_clean; i++) {
         if (cLabels[i] > num_clusters) {
           num_clusters = cLabels[i];
@@ -156,7 +156,7 @@ void MixtureModelPWClusters::run(char * criterion, int max_clusters) {
       if (num_clusters < lowest_cluster_num) {
         lowest_cluster_num = num_clusters;
         if (this->labels == NULL) {
-          this->labels = (int64_t *) malloc(sizeof(int64_t) * this->pwset->n_clean);
+          this->labels = (int *) malloc(sizeof(int) * this->pwset->n_clean);
         }
         for (int i = 0; i < this->pwset->n_clean; i++) {
           this->labels[i] = cLabels[i];
